Debug: log erreur cache brief + ajout contactEmail au lien Revoir

+++ app/brief/[id]/BriefClient.tsx
@@ -231,7 +231,7 @@ function callScoreCls(score: number) {
 }
 
 function formatCallDate(iso: string) {
-  return new Date(iso).toLocaleDateString("fr-FR", { day: "numeric", month: "short", year: "numeric" });
+  return new Date(iso).toLocaleDateString("fr-FR", { day: "numeric", month: "long", year: "numeric" });
 }
 
 export default function BriefClient({
